test: address Surfpool review cleanups
